feat: refactor DropArea component by extracting file handling and highlight logic into utility functions
