+++ app/src/Components/DropArea/DropArea.jsx
@@ -1,62 +1,11 @@
 import React, { useState } from 'react';
-import { translateInstructionToMIPS } from '../../utils/TranslatorFunctions';
+import { preventDefaults } from './utils/preventDefaults';
+import { handleDrop } from './utils/handleDrop';
+
 import './DropArea.css';
 
 const DropArea = ({ setMipsInput, setHexInput }) => {
   const [isHighlight, setIsHighlight] = useState(false);
-
-  const hexToBinary = (hex) => {
-    return parseInt(hex, 16).toString(2).padStart(8, '0');
-  };
-
-  const preventDefaults = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-  };
-
-  const handleDrop = (e) => {
-    preventDefaults(e);
-    setIsHighlight(false);
-
-    const files = e.dataTransfer.files;
-    if (files.length === 0) {
-      console.error('No files dropped');
-      return;
-    }
-
-    console.log('File dropped:', files[0]);
-
-    const reader = new FileReader();
-    reader.onload = (event) => {
-      const text = event.target.result;
-      console.log('File content:', text);
-
-      const lines = text.split('\n');
-      if (lines.length < 2) {
-        console.error('File does not contain enough lines');
-        return;
-      }
-
-      const instructionsArray = lines[1].trim().split(/\s+/);
-      let translatedInstructions = '';
-      let originalInstructions = '';
-
-      instructionsArray.forEach(instruction => {
-        const translated = translateInstructionToMIPS(instruction.trim());
-        translatedInstructions += `${translated}\n`;
-        originalInstructions += `${instruction.trim()}\n`;
-      });
-
-      setMipsInput(translatedInstructions.trim());
-      setHexInput(originalInstructions.trim());
-    };
-
-    reader.onerror = (error) => {
-      console.error('Error reading file:', error);
-    };
-
-    reader.readAsText(files[0]);
-  };
 
   return (
     <div
@@ -64,7 +13,7 @@ const DropArea = ({ setMipsInput, setHexInput }) => {
       onDragEnter={(e) => { preventDefaults(e); setIsHighlight(true); }}
       onDragOver={preventDefaults}
       onDragLeave={(e) => { preventDefaults(e); setIsHighlight(false); }}
-      onDrop={handleDrop}
+      onDrop={(e) => handleDrop(e, setIsHighlight, setMipsInput, setHexInput)}
       onMouseEnter={() => document.getElementById('dropArea').style.backgroundColor = '#9e6868'}
       onMouseLeave={() => document.getElementById('dropArea').style.backgroundColor = ''}
       className={`${isHighlight ? 'highlight' : ''} drop-area`}
