+++ app/src/Components/DropArea/utils/preventDefaults.js
@@ -1,0 +1,4 @@
+export const preventDefaults = (e) => {
+    e.preventDefault();
+    e.stopPropagation();
+  };
