+++ app/src/Components/DropArea/utils/setIsHighlight.js
@@ -1,0 +1,3 @@
+import { useState } from 'react';
+
+export const [isHighlight, setIsHighlight] = useState(false);
